Split MMIO/IIGS types into clem_mmio_types.h
