+++ clem_device.h
@@ -1,7 +1,7 @@
 #ifndef CLEM_DEVICE_H
 #define CLEM_DEVICE_H
 
-#include "clem_types.h"
+#include "clem_mmio_types.h"
 
 /**
  * Device Interface
